LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-007-expected.txt


FAIL Should use the last fallback position initially assert_equals: Anchored element should be at the top of anchor expected 600 but got 100
FAIL Should use the third fallback position with enough space left assert_equals: Anchored element should be at the left of anchor expected 135 but got 785
FAIL Should use the second fallback position with enough space below assert_equals: Anchored element should be at the right of anchor expected 135 but got 685
FAIL Should use the first fallback position with enough space left and below assert_equals: Anchored element should be at the left of anchor expected 135 but got 785
